revision

+++ kernel/keyboard.h
@@ -177,7 +177,7 @@ extern "C"  __declspec(dllexport) void __kKeyboardProc();
 extern "C"  __declspec(dllexport) void __kPutKbd(unsigned char s,int wid);
 #else
 
-extern "C" __declspec(dllimport) WORD gKeyboardID;
+extern "C" __declspec(dllimport) int gKeyboardID;
 
 extern "C"  __declspec(dllimport) DWORD gKbdTest;
 extern "C"  __declspec(dllimport) unsigned int __kGetKbd(int wid);
